requirements.txt
openai-whisper>=20231101
faster-whisper
torch>=2.2.0
torchaudio>=2.2.0
huggingface_hub
pyannote.audio>=3.1



